feat(macos): the shared SwiftUI sources build for iOS
Not an iOS app — the scene root, the split view and the menu bar stay macOS.
What crosses is everything else: 51 of the app's 60 files, after ~139 lines.

NSImage becomes a PlatformImage typealias, which the art pipeline barely
notices — it is CGImageSource end to end and only meets a platform image at
the last step. The pasteboard and the accent colour get a UIKit half, the
search field gets a plain TextField twin, and the modifiers iOS lacks are
fenced.

Two things move because portable views depend on them and they were living
in the shell: `onStage`, read by QueueView and PlayingIndicator, and AppState
itself, which is the same on either platform bar two members.

The macOS app is unchanged.

Part of #344.
Closes #347.
